Yair Levy Cruz Rios, programas con menu

+++ YLCRMenus/MainD.java
@@ -29,14 +29,14 @@ public class MainD { // Clase MainD para probar la operación de división
 
                 // Verificar división por cero y calcular la división de ambos objetos
                 if (division1.getNumero2() != 0) { // Si el divisor no es cero
-                    int resultado1 = division1.getNumero1() / division1.getNumero2(); // División del primer objeto
+                    double resultado1 = (double) division1.getNumero1() / division1.getNumero2(); // División del primer objeto
                     System.out.println("El resultado de la división del objeto 1 es: " + resultado1); // Imprimir resultado 1
                 } else { // Si el divisor es cero
                     System.out.println("Error: No se puede dividir por cero en el objeto 1."); // Mensaje de error
                 }
 
                 if (division2.getNumero2() != 0) { // Si el divisor no es cero
-                    int resultado2 = division2.getNumero1() / division2.getNumero2(); // División del segundo objeto
+                    double resultado2 = (double) division2.getNumero1() / division2.getNumero2(); // División del segundo objeto
                     System.out.println("El resultado de la división del objeto 2 es: " + resultado2); // Imprimir resultado 2
                 } else { // Si el divisor es cero
                     System.out.println("Error: No se puede dividir por cero en el objeto 2."); // Mensaje de error
